+++ src/Pages/Proposal/Components/index.js
@@ -4,3 +4,6 @@ export { default as ProposalTimingProgressBar } from './ProposalTimingProgressBa
 export { default as ProposalTimingTable } from './ProposalTimingTable';
 export { default as ProposalVotingGraph } from './ProposalVotingGraph';
 export { default as ProposalVotingTable } from './ProposalVotingTable';
+export { default as ProposalVoting } from './ProposalVoting';
+export { default as ManageVotingModal } from './ManageVotingModal';
+export { default as VotingChart } from './VotingChart';