+++ src/consts/index.js
@@ -5,5 +5,6 @@ export * from './options';
 export * from './paths';
 export * from './polling';
 export * from './staking';
+export * from './voting';
 export * from './urls';
 export * from './skips';