+++ src/redux/hooks/useTimeout.ts
@@ -1,0 +1,25 @@
+import { useEffect, useLayoutEffect, useRef } from 'react';
+
+const useTimeout = (callback: () => void, delay: number) => {
+  const savedCallback = useRef(callback);
+
+  // Remember the latest callback if it changes.
+  useLayoutEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the timeout.
+  useEffect(() => {
+    // Don't schedule if no delay is specified.
+    // Note: 0 is a valid value for delay.
+    if (!delay && delay !== 0) {
+      return undefined;
+    }
+
+    const id = setTimeout(() => savedCallback.current(), delay);
+
+    return () => clearTimeout(id);
+  }, [delay]);
+};
+
+export default useTimeout;